requirements.txt
asn1crypto==0.24.0
attrs==17.4.0
Automat==0.6.0
backports.ssl-match-hostname==3.5.0.1
BeautifulSoup==3.2.1
beautifulsoup4==4.6.0
boto3==1.4.4
certifi==2018.1.18
cffi==1.11.5
chardet==3.0.4
click==6.7
constantly==15.1.0
cryptography==2.2.1
cssselect==1.0.3
Django==1.11.11
docker-py==1.10.6
docker-pycreds==0.2.2
docutils==0.14
enum34==1.1.6
git+https://github.com/pythonforfacebook/facebook-sdk.git@449f56f0db086a41bedd23df714e7f77c1051f5b#egg=facebook_sdk
flickrapi==2.3
foursquare==20130707
futures==3.2.0
gevent==1.0.2
greenlet==0.4.7
grequests==0.2.0
httplib2==0.10.3
hyperlink==18.0.0
idna==2.6
incremental==17.5.0
ipaddress==1.0.19
iso8601==0.1.10
jmespath==0.9.3
lxml==4.2.0
oauth2==1.9.0.post1
oauthlib==2.0.2
parsel==1.4.0
#pkg-resources==0.0.0
poster==0.8.1
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.18
PyDispatcher==2.0.5
pyOpenSSL==17.5.0
python-dateutil==2.6.1
python-google-places==1.0.0
python-yelp-v2==0.5.7
pytz==2018.3
PyYAML==3.12
queuelib==1.5.0
requests==2.18.4
requests-oauthlib==0.8.0
requests-toolbelt==0.8.0
retrying==1.3.3
s3transfer==0.1.13
scrapinghub==2.0.3
Scrapy==1.4.0
service-identity==17.0.0
shub==2.7.0
six==1.11.0
tqdm==4.19.8
Twisted==17.9.0
Unidecode==0.04.16
urllib3==1.22
w3lib==1.19.0
websocket-client==0.47.0
wikipedia==1.4.0
yelp==1.0.2
yelpapi==2.0.1
zope.interface==4.4.3
